+++ android/app/src/main/java/org/threeho/summersolstice2026/MainActivity.java
@@ -1,0 +1,5 @@
+package org.threeho.summersolstice2026;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}