+++ eslint.config.mjs
@@ -5,7 +5,7 @@ const compat = new FlatCompat({ baseDirectory: import.meta.dirname });
 
 const eslintConfig = defineConfig([
   ...compat.extends("next/core-web-vitals", "next/typescript"),
-  globalIgnores([".next/**", "out/**", "build/**", "next-env.d.ts", "public/sw.js", "public/workbox-*.js", "public/swe-worker-*.js"]),
+  globalIgnores([".next/**", "out/**", "build/**", "next-env.d.ts", "public/sw.js", "public/workbox-*.js", "public/swe-worker-*.js", "android/**", "ios/**"]),
 ]);
 
 export default eslintConfig;